requirements.txt
flask
pandas
scikit-learn

nltk
numpy
requests
google-generativeai
python-dotenv
